feat: P1+P2 diagnostics — probe decomposition, obligation table, issue lifecycle (Refs PMAT-077)
P1 implementations:
- Probe-level score decomposition: per-equation/obligation drill-down
  for all 5 dimensions (D1-D5) with pass/fail probes
- Per-obligation verification table: `pv proof-status --table` shows
  L2/L3/L4 matrix per obligation (Kani property table pattern)

P2 implementation:
- Issue lifecycle: findings marked [NEW] vs pre-existing via fingerprint
  tracking in .pv/lint-previous.json (SonarQube lifecycle pattern)

Refactoring:
- Extract obligation_matrix.rs from proof_status.rs (under 500 lines)
- Split scoring_probe_tests.rs from scoring_tests.rs
- 6/13 diagnostic gaps now implemented (3 P0 + 2 P1 + 1 P2)
