+++ crates/provable-contracts/src/lib.rs
@@ -48,6 +48,7 @@ pub mod latex;
 pub mod lean_gen;
 pub mod lint;
 pub mod mirai_gen;
+pub mod obligation_matrix;
 pub mod probar_gen;
 pub mod proof_status;
 pub mod query;
